+++ src/components/recipe-form/RecipeForm.module.css
@@ -1,0 +1,134 @@
+.DialogContent {
+  display: grid;
+  gap: 1rem;
+  padding: 1rem;
+  background-color: white;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  min-width: 281px;
+  /* min px for no x-scroll */
+}
+
+.dropdownMenuButton {
+  justify-self: right;
+}
+
+.dropdownMenuContent {
+  display: grid;
+  gap: 1rem;
+  background-color: white;
+  padding: 0.5rem;
+  box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
+}
+
+.dropdownMenuItem {
+  min-width: 100px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: transparent;
+  border: none;
+}
+
+.dropdownMenuItem:hover {
+  background-color: gray;
+}
+
+.recipeTitle,
+.actionButtons {
+  justify-self: center;
+}
+
+.ingredientsContainer {
+  max-height: 288px;
+  scrollbar-width: thin;
+  overflow-y: auto;
+}
+
+.ingredientInfo {
+  position: relative;
+  padding: 1rem;
+  margin-bottom: 1rem;
+}
+
+.ingredientInfo>* input[type="number"],
+.ingredientInfo>* select {
+  width: 7ch;
+}
+
+.ingredientInfo>* input[name="name"] {
+  max-width: calc(3 * 7ch + 32px);
+}
+
+.labelInput {
+  display: grid;
+}
+
+.quantityRow,
+.macrosRow,
+.actionButtons {
+  display: flex;
+  gap: 0 1rem;
+}
+
+.deleteIngredientButton {
+  position: absolute;
+  display: inline-flex;
+  top: 6px;
+  right: 16px;
+}
+
+.AccordianRoot {
+  justify-self: start;
+}
+
+.AccordionTrigger {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  background: transparent;
+  border: none;
+}
+
+.AccordianContent {
+  font-family: Helvetica;
+}
+
+.AccordionTrigger[data-state="open"]>.AccordionChevron {
+  transform: rotate(180deg);
+}
+
+.calories {
+  font-size: 1.25rem;
+}
+
+.caloriesTotal {
+  font-size: 1.75rem;
+}
+
+.bold {
+  font-weight: bold;
+}
+
+.SeparatorRoot {
+  background-color: black;
+}
+
+.SeparatorRoot[data-orientation="horizontal"] {
+  height: 1px;
+  width: 100%;
+}
+
+.SeparatorRoot[data-orientation="vertical"] {
+  height: 100%;
+  width: 1px;
+}
+
+.newIngredientButton {
+  justify-self: left;
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
+}
